rendering board from game state returned from backend
